+++ domain/src/main/java/com/aqube/ram/domain/interactor/GetCharacterByIdUseCase.kt
@@ -5,11 +5,11 @@ import com.aqube.ram.domain.repository.CharacterRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
-typealias GetCharacterByIdBaseUseCase = BaseUseCase<Int, Flow<Character>>
+typealias GetCharacterByIdBaseUseCase = BaseUseCase<Long, Flow<Character>>
 
 class GetCharacterByIdUseCase @Inject constructor(
     private val characterRepository: CharacterRepository
 ) : GetCharacterByIdBaseUseCase {
 
-    override suspend operator fun invoke(params: Int) = characterRepository.getCharacter(params)
+    override suspend operator fun invoke(params: Long) = characterRepository.getCharacter(params)
 }
